Add GetTenantID that uses an id->service map
Prior to this change it was impossible to call MigrateServices which
specified that a hierarchical service definition be deployed.  It would
fail the validateServiceMigration while checking for conflicting
endpoints.  It failed because the check needs to start at the root
(i.e., tenant) service, which is discovered by repeatedly following
parent links.  If the parent of a service has not yet been created, (as
in the case of a service which has not yet been deployed) we cannot
follow the links to the root.

This change adds the ability to specify a dictionary mapping from ID to
service when looking up the tenant.The dictionary is pre-populated with
the not-yet-created services, allowing the determination of the proper
service root.